+++ src/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Navi } from './components/navi';
 import { Banner } from './components/banner';
+import { About } from './components/about';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './styles/main.scss'
 
@@ -10,5 +11,6 @@ root.render(
     <>
         <Navi />
         <Banner />
+        <About />
     </>
 );